pyproject.toml
[tool.poetry]
name = "tgfront"
version = "0.1.0"
description = "Telegram front-end for Health AI chatbot"
authors = ["Timofei Ryko <84790124+timofeiryko@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.8.0"
emoji-country-flag = "^1.3.2"
retry = "^0.9.2"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
